translator-service/requirements.txt
Flask==3.0.0
pytest==7.4.0
openai==1.70.0
python-dotenv==1.1.0
mock==5.2.0
